Resolve comments: renamed some files and refactored CSSes
Removed redundant nouns from the services names, removed unnecessary quotes from package.json, moved CSS overrides to themes.ts
