+++ apps/super-app/src/common/Theme.ts
@@ -37,6 +37,19 @@ const callHomeTheme = createTheme(
         fontFamily: 'Hind',
         lineHeight: '28px',
       },
+      h6: {
+        fontSize: 16,
+        fontWeight: 600,
+        color: '#297B94',
+        fontFamily: 'Hind',
+        textDecoration: 'underline',
+      },
+      subtitle1: {
+        fontSize: 14,
+        fontWeight: 400,
+        color: '#000000',
+        fontFamily: 'Hind',
+      },
     },
     palette: {
       background: {
