+++ apps/super-app/src/services/SupportServices.ts
@@ -1,4 +1,4 @@
-export interface SupportAbout {
+export interface Support {
   // Url of logo image.
   logo: string;
   // The header of the about section.
